Add the first real Foundry test: a fork test against live Aave v3
contracts/test/ was empty and had never had a test file in git history,
yet .github/workflows/test.yml ran `forge test` and exited 0 on an empty
suite -- a green badge over zero coverage.

Step 1 -- interface seam (the compile-break was real: LiquidationShield.sol
imported and typed POOL as MockPool, calling aTokenFor(), userDebt(),
userCollateral(), assetPriceUsd() -- none of which exist on real Aave's
IPool):

- New IPoolLike.sol: the real IPool subset the shield actually calls.
  getReserveData's ReserveDataLegacy struct layout was verified
  field-by-field via `cast call` against the live deployed Pool
  (0x8787...FA4E2, block 25,855,209) before being written here, not
  copied from a blog post -- same discipline as
  agent/src/reader/abis.ts's own header comment.
- New IPriceOracleLike.sol: price lives on a separate AaveOracle
  contract on real Aave, not a pool method.
- New AaveAdapter.sol: translates "aToken for X" / "user's debt in X"
  into getReserveData(X).aTokenAddress / .variableDebtTokenAddress +
  balanceOf, since neither is a direct pool method on real Aave.
- MockPool now explicitly implements both interfaces and doubles as its
  own price oracle for the mock/demo path. Added MockDebtToken, a thin
  balanceOf proxy onto MockPool's own userDebt mapping -- the mock
  previously had no debt *token* at all, just a mapping, unlike aTokens
  which already had a real (mock) token contract.
- AaveFlashProvider and HealthMath moved off MockPool-typed casts onto
  IPoolLike. Also removed the dead second return value from
  _snapshotUserState (userCollateral's result was always discarded by
  its only caller) rather than inventing an adapter for something
  nothing used.

Step 2 & 3 -- contracts/test/LiquidationShieldFork.t.sol, a forge fork
test against real Aave v3 mainnet at a pinned block (foundry.toml's
existing `mainnet` rpc_endpoint, previously declared and consumed by
nothing):

- Forks mainnet, creates a fresh WETH/USDC borrower position (not a
  live whale -- more reproducible across time), crashes WETH's oracle
  price -10% by mirroring scripts/crashPrice.ts's exact mechanism
  (MockAggregator + impersonated ACL admin + setAssetSources).
- Positive test asserts against Aave's OWN getUserAccountData, not our
  math, that HF strictly improved (1.075 -> 1.285, independently
  hand-verified beforehand and matched almost exactly) and reached
  policy.targetHF.
- Negative test: an intentionally-too-tight maxAmountIn reverts with
  SlippageExceeded (confirmed via trace it happens inside MockSwapper
  and propagates cleanly through the real Aave flash-loan callback
  chain), then asserts collateral, debt, HF, and both parties' token
  balances are byte-for-byte unchanged -- the "atomic or nothing" claim,
  previously unproven.
- Caught a real bug while building this: executeIntervention's own
  step 11 (setLastExecutedAt) is called by the shield contract itself,
  so PolicyRegistry needs the shield registered as a keeper too, not
  just the calling EOA -- the negative test didn't expose this since it
  reverts earlier; only surfaced once the happy path ran to completion.

Note why a mock-only suite could never have caught any of this:
MockPool.withdraw never decremented userCollateral, so releasing
collateral was invisible to getUserAccountData's health factor and HF
could only ever go up. Every HF invariant would have passed trivially
regardless of whether executeOperation was correct.

CI: added a guard step (forge test --list --json, fails loudly on zero
discovered tests) before the actual test step, added --fail-fast, and
wired MAINNET_RPC via secrets.MAINNET_RPC for the fork test -- needs to
be added as a repo secret for CI to actually run it. Also ran
`forge fmt` project-wide: forge fmt --check was already failing on 7
files this change never touched (pre-existing drift), which would have
made CI red again for an unrelated, confusing reason right after fixing
the empty-suite issue.

Verified: forge build --force clean, forge fmt --check clean, both fork
tests pass against live mainnet with independently hand-verified
numbers, pnpm test unaffected (91/95, same pre-existing baseline).

+++ contracts/src/PolicyRegistry.sol
@@ -63,25 +63,31 @@ contract PolicyRegistry is Ownable, IPolicyRegistry {
         lastExecutedAt[user] = timestamp;
     }
 
-    function validatePolicy(
-        address user,
-        address collateralAsset,
-        address debtAsset,
-        uint256 releaseAmount
-    ) external view override returns (Policy memory) {
+    function validatePolicy(address user, address collateralAsset, address debtAsset, uint256 releaseAmount)
+        external
+        view
+        override
+        returns (Policy memory)
+    {
         Policy memory p = _policies[user];
         if (!p.enabled) revert PolicyDisabled();
         if (block.timestamp > p.expiry) revert PolicyExpired();
 
         bool collateralOk = false;
-        for (uint i = 0; i < p.allowedCollateral.length; i++) {
-            if (p.allowedCollateral[i] == collateralAsset) { collateralOk = true; break; }
+        for (uint256 i = 0; i < p.allowedCollateral.length; i++) {
+            if (p.allowedCollateral[i] == collateralAsset) {
+                collateralOk = true;
+                break;
+            }
         }
         if (!collateralOk) revert AssetNotAllowed();
 
         bool debtOk = false;
-        for (uint i = 0; i < p.allowedDebt.length; i++) {
-            if (p.allowedDebt[i] == debtAsset) { debtOk = true; break; }
+        for (uint256 i = 0; i < p.allowedDebt.length; i++) {
+            if (p.allowedDebt[i] == debtAsset) {
+                debtOk = true;
+                break;
+            }
         }
         if (!debtOk) revert AssetNotAllowed();
